wire operator metrics through api monitoring
